chore: remove etherscan support
