--- src/etherscan/enums/etherscan-action.enum.ts
@@ -1,3 +1,0 @@
-export enum EtherscanAction {
-  GetBlockNoByTime = 'getblocknobytime',
-}
