--- src/etherscan/enums/etherscan-module.enum.ts
@@ -1,3 +1,0 @@
-export enum EtherscanModule {
-  Block = 'block',
-}
