--- src/etherscan/interfaces/etherscan-block-response.interface.ts
@@ -1,5 +1,0 @@
-export interface EtherscanBlockResponse {
-  status: string;
-  message: string;
-  result: string;
-}
